moves rootSaga folder into the redux dir
